+++ Source/SpoutMediaIO/Private/SpoutMediaIOModule.cpp
@@ -1,0 +1,5 @@
+#include "SpoutMediaIOModule.h"
+
+#include "Modules/ModuleManager.h"
+
+IMPLEMENT_MODULE(FSpoutMediaIOModule, SpoutMediaIO)