+++ Source/SpoutMediaIO/Public/SpoutMediaIOModule.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Modules/ModuleInterface.h"
+
+class FSpoutMediaIOModule : public IModuleInterface
+{
+};